[android][downloader] download completed service is unified.
