+++ android/AndroidManifest.xml
@@ -465,7 +465,7 @@
     </receiver>
 
     <receiver
-      android:name="com.mapswithme.maps.bookmarks.SystemDownloadCompletedReceiver"
+      android:name="com.mapswithme.maps.background.SystemDownloadCompletedReceiver"
       android:exported="true">
       <intent-filter>
         <action android:name="android.intent.action.DOWNLOAD_COMPLETE"/>
@@ -473,7 +473,7 @@
     </receiver>
 
     <service
-      android:name="com.mapswithme.maps.bookmarks.SystemDownloadCompletedService"
+      android:name="com.mapswithme.maps.background.SystemDownloadCompletedService"
       android:permission="android.permission.BIND_JOB_SERVICE"
       android:exported="false">
     </service>
