+++ android/src/com/mapswithme/maps/scheduling/JobIdMap.java
@@ -2,7 +2,7 @@ package com.mapswithme.maps.scheduling;
 
 import com.mapswithme.maps.background.NotificationService;
 import com.mapswithme.maps.background.WorkerService;
-import com.mapswithme.maps.bookmarks.SystemDownloadCompletedService;
+import com.mapswithme.maps.background.SystemDownloadCompletedService;
 import com.mapswithme.maps.geofence.GeofenceTransitionsIntentService;
 import com.mapswithme.maps.location.TrackRecorderWakeService;
 import com.mapswithme.util.Utils;
